Added Method to Help Enforce Fresh Buffer Each Frame
Added a method to the PixyCamWrapper that lets the user of the wrapper
to force the wrapper to refresh the internal buffer for each frame.
Using this method instead of having the user of the wrapper call
`GetByteArray()` or something of that nature allows for multiple
interfaces to use different methods of sending bytes.  For example if
SPI ever works, that uses big endian while I2C uses little endian.